Improve resiliency
This commit fixes issue #1.

The initial goal of this commit was to fix issue #1. For this, this
commit changes the job's name to be created automatically by the
APIserver and thus keeps track of the jobname through an annotation.
While testing this change, I ran into a panic somewhere in the code.
With this panic, I went down a rabbit hole of "issues that may appear in
edge cases".
I decided to change the way the phases where setup. So far, the phases
where responsible to check if the state was populated with the objects
they needed. However, no phase actually really checked that, so that's
where the panics came from. This has been solved by moving the whole
setup and checking of the state into the `newState` function, which now
maps the objects depending on the phase that the PVC is in.

Because of these checks, the tool is expected to exit earlier in case of
a weird state. It is also more specific on the "requirements" of every
phase.

Further, the annotations have been renamed and are now prefixed with
`relokator` (so `relokator.nine.ch/...`). This is to improve clarity.

Another issue that is fixed with this commit is that the annotations map
may be nil when trying to add an annotation. This has been fixed by
extracting the work that is needed (e.g. nil-check and make) into an own
function that can now be used to add annotations.

The clusterState has been updated to use only a string (e.g. the name of
the namespace) instead of the full namespace. This makes setting up a
bit easier. Same for the namespaceState.

Also fixes the dockerfile and adds a github action.

Signed-off-by: Ramon Rüttimann <ramon@nine.ch>