Fix finding PVCs generated by StatefulSets
StatefulSets were somewhat broken before this. Usually a StatefulSet
uses PersistentVolumeTemplates to dynamically allocate PVCs depending
on the number of replicas. When trying to migrate such a volume, relokator
did not find the associated StatefulSet, as it only looked in the volumes
list of the pod template, which is usually empty as it will be filled by the
StatefulSet controller. We still look for volumes in the template now, as this
could also be a valid use case (e.g. for RWX volumes that every replica mounts).